requirements.txt
streamlit==1.53.1
ultralytics==8.4.11
pillow==12.1.0
opencv-python-headless==4.13.0.90
